Modularization pt 2
