+++ internal/app/server/server.go
@@ -6,10 +6,10 @@ import (
 	"net/http"
 	"strconv"
 
-	"github.com/egert811/task-server/internal/pkg"
 	"time"
 	"fmt"
 	"context"
+	"github.com/egert811/task-server/internal/pkg/storage"
 )
 
 // server config, TODO: externalize
