+++ internal/app/worker/worker.go
@@ -1,0 +1,2 @@
+package worker
+
