+++ internal/pkg/storage/model.go
@@ -1,0 +1,11 @@
+package storage
+
+type TaskDBItem struct {
+	ID  int    `json:"id"`
+	CMD string `json:"cmd"`
+}
+
+type TaskOutputDBItem struct {
+	ID     int    `json:"-"`
+	Output string `json:"output"`
+}
